Add 64-bit compat stubs and fix prototypes for cross-platform build
- Add minimal dos.h, conio.h, bios.h stubs for 64-bit native builds
- Fix function prototypes and argument types for cross-platform
  compatibility
- Use extern for global variables referenced from multiple modules
- Update slot function signatures for BUGFIX builds
- Add missing includes and macro guards for portability
- Remove unused/obsolete CMake and header includes

+++ src/shared/util.c
@@ -8,10 +8,11 @@
 #include "util.h"
 
 #include "../pointers.h"
+#include <dos.h>
 
 /* extern declarations needed by these functions */
 extern void far gfx_drawString(int16 *pageNum, const char *string);
-extern int far gfx_setFont(int ch, int font);
+extern int far gfx_setFont(uint16 ch, uint16 font);
 extern void far misc_jump_5e_clearKeyFlags(void);
 extern char timerHandlerInstalled;
 void restoreTimerIrqHandler(void);
